module2-finDistill/requirements.txt
azure-functions
azure-storage-blob
azure-search-documents
groq
pypdf2
redis
python-dotenv